Update three_roaming_scraper.py
OK, so running out of time regarding the for loop to go through the table so coded to retreive the rows reqiured. THis will work fine but I'd rather be more elegant with it. Will keep working but the info required is being returned to the console. Would like to move on to storing data and presenting